odp: Honor Open Flow spec when converting from OF to DP flow for neighbor discovery.
The OpenFlow spec doesn't require that a user specify icmp_code when specifying a type.  However,
the conversion for a DP flow asks that the user explicitly specified an icmp_code field to match
and forces this via a mask check.  This means that valid matches for icmp_type=136,... (for
example) won't properly generate a full flow and there will be a much broader match installed
in the kernel datapath.

This can be worked around by explicitly including icmp_code, but for users that want to write
flows which are supported by the kernel, it is not possible to omit icmp_code and still have
a neighbor discovery match included.

An alternative way to fix up the flow and mask would be to modify the output of the xlate in
the xlate_wc_finish() to set the mask when detecting a neighbor discovery related packet.
This would result in further changes needed, as every flow generated would now include these
masks, and it would potentially result in flows that are constantly revalidated without even
more complex logic.  The approach taken here is to relax the requirements from the ODP side.

Signed-off-by: Aaron Conole <aconole@redhat.com>