ovs-tcpdump: Improve 'mirror-to' option.
When the 'mirror-to' option was conceived, it was intended to allow users
the ability to arbitrarily build their own ports to use the mirror port
rather than always using an algorithmically defined mirror port.  However,
the mirror port code never accommodated a port that the user defined as
part of an OVS bridge.  This could be useful for running against all kinds
of OVS specific ports.

Adjust the 'mirror-to' option so that when the user specifies a port, the
utility searches through system interfaces, as well as the OVS DB.  This
means that we need to drop the port_exists() check for the mirror port,
but that should be okay.

Reported-at: https://mail.openvswitch.org/pipermail/ovs-discuss/2025-March/053508.html
Reported-by: Jun Wang <junwang01@cestc.cn>
Signed-off-by: Aaron Conole <aconole@redhat.com>