checkpatch: Separate out the built-in spelling words.
Checkpatch contains a large list of spelling words, currently 326.  The
in-built list is quite cumbersome, and includes duplicate entries
('netdev' and 'revalidation').  Managing this list will continue to grow
difficult over time, and really shouldn't be so difficult.  To make the
naintenance of OVS specific spelling words easier, create a new file
that will contain the spelling words and move the list into that file
(sorted).  Additionally, add 'FreeBSD' and 'gcloud' to the list as I
tested against 80d723736b64 ("cirrus: Update to FreeBSD 14.3 and 13.5.")

Signed-off-by: Aaron Conole <aconole@redhat.com>